tests: split test_workspace from test_resolver
