+++ requirements_test.txt
@@ -5,3 +5,4 @@ sphinx
 codecov >= 2.0.15
 recommonmark >= 0.5.0
 docstr-coverage
+pylint
